+++ src/main/java/com/study/algorithm/seungjae/boj/Ex11399.java
@@ -1,30 +1,37 @@
-package com.study.algorithm.seungjae;
+package com.study.algorithm.seungjae.boj;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.Arrays;
+import java.util.StringTokenizer;
 
-public class Ex2217 {
+public class Ex11399 {
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
+
+        StringTokenizer st = new StringTokenizer(br.readLine());
+
         int[] arr = new int[n];
 
         for (int i = 0; i < n; i++) {
-            arr[i] = Integer.parseInt(br.readLine());
+            arr[i] = Integer.parseInt(st.nextToken());
         }
 
         Arrays.sort(arr);
 
-        int max = Integer.MIN_VALUE;
+        int count = 0;
 
         for (int i = 0; i < n; i++) {
-            max = Math.max(max, arr[i] * (n - i));
+            count += arr[i];
+            for (int j = 0; j < i; j++) {
+                count += arr[j];
+            }
         }
 
-        System.out.println(max);
+        System.out.println(count);
     }
 
 }
